chore: refactor core structure
