+++ cli/internal/cmd/add/add.go
@@ -10,9 +10,9 @@ import (
 	"path/filepath"
 
 	"github.com/aerogear/charmil/cli/internal/common/modname"
-	"github.com/aerogear/charmil/cli/internal/factory"
 	"github.com/aerogear/charmil/cli/internal/template/add"
-	"github.com/aerogear/charmil/core/color"
+	"github.com/aerogear/charmil/cli/pkg/factory"
+	"github.com/aerogear/charmil/core/utils/color"
 	"github.com/spf13/cobra"
 )
