+++ starter/cmd/cli/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/aerogear/charmil/core/config"
 	"github.com/aerogear/charmil/core/localize"
-	"github.com/aerogear/charmil/starter/internal/factory"
 	"github.com/aerogear/charmil/starter/internal/update"
+	"github.com/aerogear/charmil/starter/pkg/factory"
 	"github.com/spf13/cobra"
 	"github.com/spf13/cobra/doc"
 	"golang.org/x/text/language"
