+++ starter/pkg/cmd/completion/bash/bash.go
@@ -1,7 +1,7 @@
 package bash
 
 import (
-	"github.com/aerogear/charmil/starter/internal/factory"
+	"github.com/aerogear/charmil/starter/pkg/factory"
 	"github.com/spf13/cobra"
 )
