+++ starter/pkg/cmd/completion/fish/fish.go
@@ -1,7 +1,7 @@
 package fish
 
 import (
-	"github.com/aerogear/charmil/starter/internal/factory"
+	"github.com/aerogear/charmil/starter/pkg/factory"
 	"github.com/spf13/cobra"
 )
