+++ starter/pkg/cmd/completion/zsh/zsh.go
@@ -1,7 +1,7 @@
 package zsh
 
 import (
-	"github.com/aerogear/charmil/starter/internal/factory"
+	"github.com/aerogear/charmil/starter/pkg/factory"
 	"github.com/spf13/cobra"
 )
